RWA-968 refactor
